Move all functions to a module (fortran)
